refactor(auth): update imports to use schemas/auth instead of types/auth
- Update constants/auth.ts to import User from schemas/auth
- Update contexts/AuthContext.tsx to import User and AuthContextType from schemas/auth
- Update hooks/useAuth.ts to import AuthContextType from schemas/auth
- Update ProtectedRoute test to use schemas/auth
