+++ packages/ui/src/hooks/useAuth.ts
@@ -3,7 +3,7 @@
  */
 import { useContext } from 'react';
 import { AuthContext } from '../contexts/AuthContext';
-import type { AuthContextType } from '../types/auth';
+import type { AuthContextType } from '../schemas/auth';
 
 export function useAuth(): AuthContextType {
   const context = useContext(AuthContext);
